fix post endpoints
