+++ src/routes/user.route.js
@@ -4,5 +4,5 @@ import userController from "../controllers/user.controller.js";
 
 const userRoute = Router();
 
-userRoute.get('/', userController);
+userRoute.get('/', userController);  
 export default userRoute;
